Uniform seed respresentation in the API.
Subject seeds and other seeds are represented differently in the list object. That makes the API ugly. This commit fixes that by providing uniform seeds interface.

In the list object, seeds are represented like this:

    {"key": "/books/OL1M"}
    "subject:love"

Now, they are exposed in the API as:

    "/books/OL1M"
    "/subjects/love"

Eventually, the seed representation in the list object should also be changed to this.