Save book borrow status in the store and update it on every borrow and return.
To find whether or not a book is borrowed requires a query, which is not
cachable. To avoid that query, an entry for each ebook is created in the
store and its status updated on every borrow and return. Since there is
one entry for each book and borrow info can be fetched and cached
easily.