Implement proper deletion and flagging of emails.
We fetch all "New" messages and go through them. If a message Subject
matches the regexp for our support system, try to update the support
database. If successful, mark the message for deletion. If not, reset
the message to "Unseen". If the message subject didn't match our
support regexp, mark it as unseen. Finally, expunge all deleted
messages. The code to update the database is not in this commit.