Randomized the display of books in /borrow. #85
Solr can sort the results in random if we ask to sort by "random_1234
desc". We can pass any number instead of 1234 and it always returns the
same order when we use the same number.

Randomized using sort by random and passed the random number to the
template so that it can be used when making the ajax requests for
pagination.