Improved solr-updater to avoid one query. #104
The query_iter function was querying OL in a loop with increasing offset
until it finds no results. The teriminating condition can be detected
one step earlier by checking size of the response. If the size of the
response is less than the limit, then it is the last chunk.

This saves one call to query each time query_iter is called.