Delete ia:foo00bar keys when indexing a work.
Some archive.org text items are missing in OL. To overcome that we've
added fake edition records to OL that'll display the data from
archive.org metadata. These fake records are also indexed in Solr so
that they are discoverable via search engine.

To avoid duplicate entries when a fake record gets converted to real
record, all the ia:foo00bar entries are deleted from Solr when a work is
indexed.